+++ app/main/forms.py
@@ -1,10 +1,12 @@
 import wtforms as wtf
 import wtforms.widgets.html5 as widgets
 from flask_wtf import FlaskForm
+from flask_login import current_user
 from wtforms.validators import (DataRequired, InputRequired, NumberRange,
                                 Optional)
 
 from .models import Product, StockProduct, Transaction, Specification
+from app.auth.models import User
 
 
 class OrderItemForm(FlaskForm):
@@ -83,8 +85,10 @@ class ConsumeProductForm(FlaskForm):
                  | Lote: {sp.lot_number}
                  | Validade: {sp.expiration_date.strftime('%d-%m-%y')}
                  | Quantidade: {sp.amount}
+                 | Fabricante: {sp.manufacturer}
             ''')) for sp in stock_products
         ]
+        self.consumer_id.choices = [(user.id, user.email) for user in User.query.order_by(User.email).all()]
 
     stock_product_id = wtf.SelectField(
         'Reativo', coerce=int, validators=[InputRequired()])
@@ -98,6 +102,11 @@ class ConsumeProductForm(FlaskForm):
         widget=widgets.NumberInput(),
         render_kw={'autocomplete': 'off'},
         default=1,
+    )
+    consumer_id = wtf.SelectField(
+        'Consumidor final',
+        coerce=int,
+        validators=[InputRequired()],
     )
     submit = wtf.SubmitField('Confirmar')